% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text2df-package.R
\docType{package}
\name{text2df-package}
\alias{text2df}
\alias{text2df-package}
\title{text2df: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
