% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tif2parallel.R
\name{tif2parallel}
\alias{tif2parallel}
\title{Parallel text processing and annotation -- the latter via `udpipe`.}
\usage{
tif2parallel(
  tif,
  mwe = NULL,
  output_dir = NULL,
  ud_model_dir,
  cores = 6,
  per = 1
)
}
\arguments{
\item{tif}{A dataframe}

\item{mwe}{A character vector}

\item{output_dir}{A file path}

\item{ud_model_dir}{A file path}

\item{cores}{An integer}

\item{per}{An integer}
}
\value{
A list of data frames
}
\description{
Parallel text processing and annotation -- the latter via `udpipe`.
}
