% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tif2token.R
\name{tif2token}
\alias{tif2token}
\title{Convert TIF to tokens list -- via corpus package}
\usage{
tif2token(tif)
}
\arguments{
\item{tif}{A TIF}
}
\value{
A list
}
\description{
Convert TIF to tokens list -- via corpus package
}
