% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tok2collocation.R
\name{tok2collocations}
\alias{tok2collocations}
\title{Extract collocations from token object -- via text2vec package}
\usage{
tok2collocations(
  tok,
  collocation_count_min = 15,
  pmi_min = 5,
  remove_stops = F,
  remove_punct = T,
  n_iter = 5
)
}
\arguments{
\item{collocation_count_min}{An integer}

\item{pmi_min}{An integer}

\item{remove_stops}{A Boolean}

\item{remove_punct}{A Boolean}

\item{n_iter}{An integer}

\item{tokA}{A list}
}
\value{
A list
}
\description{
Extract collocations from token object -- via text2vec package
}
