% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token2annotation.R
\name{token2annotation}
\alias{token2annotation}
\title{Combine multi-word expressions per dictionary -- via quanteda --}
\usage{
token2annotation(tok, model, tagger = "default", parser = "none")
}
\arguments{
\item{tok}{A list}

\item{model}{udpipe model}

\item{tagger}{udpipe tagger}

\item{parser}{udpipe parser}
}
\value{
A data frame
}
\description{
Combine multi-word expressions per dictionary -- via quanteda --
}
