% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token2df.R
\name{token2df}
\alias{token2df}
\title{Convert tokens to df -- via textshape package}
\usage{
token2df(tok)
}
\arguments{
\item{tok}{A list}
}
\value{
A data frame
}
\description{
Convert tokens to df -- via textshape package
}
