% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token2mwe.R
\name{token2mwe}
\alias{token2mwe}
\title{Combine multi-word expressions per dictionary -- via quanteda --}
\usage{
token2mwe(tok, mwe)
}
\arguments{
\item{tok}{A list}

\item{mwe}{A character vector}
}
\value{
A data frame
}
\description{
Combine multi-word expressions per dictionary -- via quanteda --
}
